Refactor: Remove Event and Speaker Specific Pages (#1243)
Removed techdebt - like events
